load sentence splitter

+++ shared/modelling_utils/modelling.py
@@ -472,13 +472,13 @@ class Modelling:
             # new
             logits = output.logits.detach().cpu()
             active_loss = batch['attention_mask'].view(-1) == 1
-            active_logits = logits.view(-1, len(self.args.labels))
+            active_logits = logits.view(-1, len(self.args.labels)).float()
             active_labels = torch.where(
                 active_loss, batch['labels'].view(-1),
                 torch.tensor(self.weighted_loss_function.ignore_index).type_as(
                     batch['labels'])
             )
-            loss = self.weighted_loss_function(active_logits, active_labels)
+            loss = self.weighted_loss_function(active_logits, active_labels.float())
 
         loss.backward()
         optimizer.step()
